refactor: componentize views
