+++ src/components/Button/row.module.scss
@@ -1,0 +1,7 @@
+.row {
+	display: flex;
+
+	& button:first-child {
+		margin-right: var(--horizontal-gap);
+	}
+}
