+++ src/styles/base.scss
@@ -39,7 +39,7 @@
 	--color-lightgreen: rgb(141, 221, 122);
 	--color-lavender: rgb(136, 107, 214);
 	--color-red: rgb(185, 66, 66);
-	--color-orange: rgb(185, 114, 66);
+	--color-orange: rgb(231, 172, 62);
 
 	// Building icons
 	--color-building-economic: radial-gradient(
