+++ src/views/Game/index.module.scss
@@ -1,0 +1,41 @@
+.container {
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	display: flex;
+	height: 12.5rem;
+	flex-direction: column;
+	align-items: center;
+	justify-content: flex-end;
+	transform: translate3d(-50%, -50%, 0);
+
+	// Shift view down a tad to mimic real game
+	&-grid-style {
+		top: 60% !important;
+	}
+
+	.debug & {
+		border: 1px solid blue;
+	}
+}
+
+// Handle error
+.click-anywhere {
+	position: fixed;
+	top: 0;
+	left: 0;
+	display: flex;
+	width: 100%;
+	height: 100%;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+
+	& h3 {
+		margin: 0;
+	}
+	& p {
+		margin: 0;
+		color: var(--color-font-tertiary);
+	}
+}
