+++ src/views/Game/score.module.scss
@@ -1,0 +1,3 @@
+.score {
+	opacity: 0.2;
+}
