+++ src/views/Options/gameOptions.module.scss
@@ -1,0 +1,4 @@
+.container {
+	height: 100%;
+	padding: 3rem;
+}
