+++ src/views/Result/Hint.jsx
@@ -1,0 +1,12 @@
+import styles from './hint.module.scss'
+import { MIN_BUILDING_TRESHOLD } from '.'
+
+const Hint = () => {
+	return (
+		<div className={`${styles.item} ${styles.hint}`}>
+			Stats require at least {MIN_BUILDING_TRESHOLD} played buildings.
+		</div>
+	)
+}
+
+export default Hint
