+++ src/views/Result/hint.module.scss
@@ -1,0 +1,7 @@
+.hint {
+	padding: 1rem;
+	margin-bottom: var(--horizontal-gap);
+	background: var(--color-lavender);
+	border-radius: var(--border-radius);
+	font-weight: 600;
+}
