+++ src/views/Result/index.jsx
@@ -1,59 +1,76 @@
 /**
- * @file Shows some stats after the game has finished
+ * @file ...
  */
 
-import styles from './result.module.scss'
-import useStore from '../store/store'
-
-import buildings from '../assets/buildings'
+import useStore from '../../store'
+import { roundToTwoDecimals } from '../../utils/math'
+import styles from './index.module.scss'
+import Hint from './Hint'
+import Button from '../../components/Button'
+import ButtonRow from '../../components/Button/Row'
+import buildings from '../../assets/buildings'
 
 const MIN_BUILDING_TRESHOLD = 15
-const roundToTwoDecimals = (value) => Math.round(value * 100) / 100
 
-const ResultView = () => {
-	const { score, endResult } = useStore((state) => ({
-		score: state.score,
-		endResult: state.endResult
-	}))
-
-	const hideStats = score < MIN_BUILDING_TRESHOLD
+const Result = () => {
+	const { setOptionsModalOpen, score, endResult, handleGameStart } = useStore(
+		(state) => ({
+			setOptionsModalOpen: state.setOptionsModalOpen,
+			score: state.score,
+			endResult: state.endResult,
+			handleGameStart: state.handleGameStart
+		})
+	)
+	const notEnoughData = score < MIN_BUILDING_TRESHOLD
 
 	return endResult ? (
 		<div className={styles.container}>
-			<h2>Your result</h2>
-			<div className={styles['result-container']}>
-				{hideStats && <Hint />}
-				<div className={styles['result-row']}>
+			<h2 style={{ alignSelf: 'flex-start' }}>Your result</h2>
+			<div className={styles.result}>
+				{notEnoughData && <Hint />}
+				<div className={styles.row}>
 					<Item
 						value={endResult[0].value}
 						type={endResult[0].type}
-						disabled={hideStats}
+						disabled={notEnoughData}
 					/>
 					<Item
-						value={hideStats ? 0 : endResult[1].value}
+						value={notEnoughData ? 0 : endResult[1].value}
 						type={endResult[1].type}
-						disabled={hideStats}
+						disabled={notEnoughData}
 					/>
 					<Item
-						value={hideStats ? 0 : endResult[2].value}
+						value={notEnoughData ? 0 : endResult[2].value}
 						type={endResult[2].type}
-						disabled={hideStats}
+						disabled={notEnoughData}
 					/>
 				</div>
-
-				<div className={styles['result-row']}>
+				<div className={styles.row}>
 					<FastestSlowestBuilding
 						value={endResult[3]?.value}
 						type={endResult[3]?.type}
-						disabled={hideStats}
+						disabled={notEnoughData}
 					/>
 					<FastestSlowestBuilding
 						value={endResult[4]?.value}
 						type={endResult[4]?.type}
-						disabled={hideStats}
+						disabled={notEnoughData}
 					/>
 				</div>
 			</div>
+			<ButtonRow>
+				<Button
+					onClick={handleGameStart}
+					style={{ minWidth: '12rem' }}
+					primary>
+					{endResult ? 'Play again' : `Start typin'`}
+				</Button>
+				<Button
+					onClick={() => setOptionsModalOpen(true)}
+					style={{ position: 'relative' }}>
+					Options
+				</Button>
+			</ButtonRow>
 		</div>
 	) : (
 		<div style={{ padding: '4rem 0' }}>Couldn't fetch end result.</div>
@@ -77,13 +94,12 @@ const FastestSlowestBuilding = ({ value, type, disabled }) => {
 	const _building = value
 		? buildings.find((el) => value[0] === el.name)
 		: undefined
-
 	const _value = value ? value[1].toFixed(2) : 0
 
-	// Each shortcut is stored as a coordinate for an 2d array, ex. 0:0
+	// Since each building just contains the shortcut as 2d coordinate,
+	// we need to get it by looking up the current picked keyboard layout.
 	const firstKeyIndex = _building ? _building.shortcut[0].split(':') : null
 	const secondKeyIndex = _building ? _building.shortcut[1].split(':') : null
-
 	const firstKey = firstKeyIndex
 		? keyMap[firstKeyIndex[0]][firstKeyIndex[1]]
 		: null
@@ -114,7 +130,6 @@ const FastestSlowestBuilding = ({ value, type, disabled }) => {
 				<h2>
 					{_value}
 					<span className={styles.unit}>sec/avg.</span>
-					{type === 'accuracy' && '%'}
 				</h2>
 				<span>{type}</span>
 			</div>
@@ -122,12 +137,4 @@ const FastestSlowestBuilding = ({ value, type, disabled }) => {
 	)
 }
 
-const Hint = () => {
-	return (
-		<div className={`${styles.item} ${styles.hint}`}>
-			Stats require at least {MIN_BUILDING_TRESHOLD} played buildings.
-		</div>
-	)
-}
-
-export default ResultView
+export { Result, MIN_BUILDING_TRESHOLD }
