--- src/views/game.module.scss
@@ -1,196 +1,0 @@
-.container {
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	display: flex;
-	height: 12.5rem;
-	flex-direction: column;
-	align-items: center;
-	justify-content: flex-end;
-	transform: translate3d(-50%, -50%, 0);
-
-	// Shift view down a tad to mimic real game
-	&-grid-style {
-		top: 60% !important;
-	}
-
-	.debug & {
-		border: 1px solid blue;
-	}
-}
-
-.score {
-	opacity: 0.2;
-}
-
-/*
- * Single icon style
- */
-.icon {
-	width: 100%;
-	height: 100%;
-
-	&-container {
-		position: absolute;
-		z-index: 10;
-		top: 0;
-		left: 50%;
-		width: 4rem;
-		height: 4rem;
-		padding: 0.375rem;
-		border-radius: var(--border-radius-sm);
-		transform: translateX(-50%);
-
-		&.economic {
-			background: var(--color-building-economic);
-		}
-		&.military {
-			background: var(--color-building-military);
-		}
-		&.fortified {
-			background: var(--color-building-fort);
-		}
-		&.research {
-			background: var(--color-building-research);
-		}
-	}
-}
-
-.name {
-	position: absolute;
-	top: 5.5rem;
-	left: 50%;
-	color: rgba(255, 255, 255, 0.2);
-	font-size: 1.25rem;
-	font-weight: 400;
-	transform: translateX(-50%);
-	white-space: nowrap;
-}
-
-/*
- * Grid icon style
- */
-.grid {
-	display: flex;
-}
-
-.ages {
-	display: grid;
-	margin-right: 1rem;
-	margin-bottom: calc(var(--vertical-gap) * 1.5);
-	grid-gap: 2px;
-}
-
-.row {
-	display: flex;
-	grid-gap: 2px;
-	grid-template-columns: 1fr 1fr 1fr 1fr;
-}
-
-.column {
-	position: relative;
-	display: block;
-	width: 4rem;
-	height: 4rem;
-	background: #000;
-	border-radius: 0.25rem;
-	opacity: 0.2;
-	transition: transform 0.2s, box-shadow 0.2s;
-
-	&.visible {
-		background: none;
-		opacity: 1;
-
-		&.economic {
-			//TODO: merge w/ class above (redundancy)
-			background: var(--color-building-economic);
-		}
-		&.military {
-			background: var(--color-building-military);
-		}
-		&.fortified {
-			background: var(--color-building-fort);
-		}
-		&.research {
-			background: var(--color-building-research);
-		}
-	}
-
-	&.scored {
-		z-index: 100;
-		box-shadow: 0 0 2rem var(--color-green) inset;
-	}
-}
-
-.keys {
-	display: flex;
-}
-
-.key {
-	display: flex;
-	width: 3.5rem;
-	height: 3.5rem;
-	align-items: center;
-	justify-content: center;
-	border: 1px solid rgba(255, 255, 255, 0.05);
-	background: linear-gradient(
-			180deg,
-			var(--color-background) 0%,
-			var(--color-foreground) 100%
-		),
-		var(--color-background);
-	border-radius: var(--border-radius);
-	box-shadow: 0rem 0.375rem 0rem rgba(0, 0, 0, 0.25);
-	color: rgba(0, 0, 0, 0);
-	font-family: var(--key-font);
-	font-size: 1.125rem;
-	font-weight: 500;
-	text-transform: uppercase;
-	transition: transform 0.1s, box-shadow 0.1s;
-	user-select: none;
-
-	&:first-child {
-		margin-right: 1rem;
-	}
-
-	&.down {
-		box-shadow: none !important;
-		transform: translateY(0.25rem);
-	}
-
-	&.correct {
-		border: 1px solid var(--color-lightgreen);
-		background: linear-gradient(
-				180deg,
-				var(--color-green) 0%,
-				var(--color-lightgreen) 100%
-			),
-			var(--color-green);
-		color: var(--color-font-primary) !important;
-	}
-
-	&.labeled {
-		color: rgba(255, 255, 255, 1) !important;
-	}
-}
-
-// Handle error
-.click-anywhere {
-	position: fixed;
-	top: 0;
-	left: 0;
-	display: flex;
-	width: 100%;
-	height: 100%;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-
-	& h3 {
-		margin: 0;
-	}
-	& p {
-		margin: 0;
-		color: var(--color-font-tertiary);
-	}
-}
